- fixed App logged in processing - removed NR_DATA after affiliation
