+++ src/mixins/index.ts
@@ -4,5 +4,6 @@ export * from './payment-session-mixin'
 export * from './common-mixin'
 export * from './date-mixin'
 export * from './displayed-component-mixin'
+export * from './load-keycloak-roles-mixin'
 export * from './nr-affiliation-mixin'
 export * from './update-user-mixin'
