+++ src/plugins/getConfig.ts
@@ -28,8 +28,8 @@ export async function getConfig (): Promise<EnvConfigI> {
   const paymentPortalUrl = response.data['PAYMENT_PORTAL_URL']
   sessionStorage.setItem('PAYMENT_PORTAL_URL', paymentPortalUrl)
 
-  const dashboadUrl = response.data['DASHBOARD_URL']
-  sessionStorage.setItem('DASHBOARD_URL', dashboadUrl)
+  const businessesUrl = response.data['BUSINESSES_URL']
+  sessionStorage.setItem('BUSINESSES_URL', businessesUrl)
 
   const keycloakConfigPath = response.data['KEYCLOAK_CONFIG_PATH']
   sessionStorage.setItem('KEYCLOAK_CONFIG_PATH', keycloakConfigPath)
